test_hf_space/requirements.txt
# Generated from pyproject.toml for Render deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart
python-dotenv
litellm>=1.60.0,!=1.64.4
aiohttp>=3.9.0,!=3.11.13
google-generativeai
termcolor
docker
toml
types-toml
numpy
json-repair
browsergym-core==0.13.3
html2text
pexpect
jinja2>=3.1.3
tenacity>=8.5,<10.0
zope-interface==7.2
pathspec>=0.12.1
pyjwt>=2.9.0
dirhash
tornado
rapidfuzz>=3.9.0
whatthepatch>=1.0.6
protobuf>=5.0.0,<6.0.0
opentelemetry-api>=1.33.1
opentelemetry-exporter-otlp-proto-grpc>=1.33.1
libtmux>=0.37,<0.40
pygithub>=2.5.0
joblib
python-socketio>=5.11.4
sse-starlette>=2.1.3
psutil
python-json-logger>=3.2.1
prompt-toolkit>=3.0.50
anyio==4.9.0
fastmcp>=2.5.2
python-frontmatter>=1.1.0
redis>=5.2,<7.0
e2b>=1.0.5,<1.6.0
google-api-python-client>=2.164.0
google-auth-httplib2
google-auth-oauthlib
modal>=0.66.26,<0.78.0
runloop-api-client==0.39.0
openhands-aci==0.3.0
poetry>=2.1.2
pythonnet
ipywidgets>=8.1.5
qtconsole>=5.6.1
PyPDF2
python-pptx
pylatexenc
python-docx
bashlex>=0.18
minio>=7.2.8
daytona-sdk==0.20.0
stripe>=11.5,<13.0
google-cloud-aiplatform
anthropic
boto3